Move email presenter modules and tests into appropriate folders
This change moves all the email presenters into their own subfolder of
arbeitszeit_web "arbeitszeit_web/email". The associated tests were
also moved into their own subfolder "tests/email_presenters".
